Remove redundancies and modularize
